pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "bilitool" # make sure your module name is unique
version = "0.1.0"
authors = [
  { name="timerring"},
]
description = "Login and upload videos to bilibili"
readme = "README.md"
license = { file="LICENSE" }
requires-python = ">=3.10"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
  "PyYAML==6.0.2",
  "qrcode==8.0",
  "Requests==2.32.3"
]

[project.scripts]
bilitool = "bilitool.cli:cli"

[project.urls]
"Homepage" = "https://github.com/timerring/bilitool"
